feat: formato .ottdmap v3 con m6/m8; gfx de industria 9 bits
Formato .ottdmap extendido a v3:
- Agrega W×H bytes de m6 (MAPE en savegame) con bit 2 = bit 8 del gfx
  de industria (GetCleanIndustryGfx usa 9 bits en OpenTTD)
- Agrega W×H×2 bytes de m8 (MAP8) con HouseID (u16 LE) para casas

Correcciones de chunk names en parse_sav.py:
- MAP1 se llama 'MAPO' en el savegame (no 'MAP1' → m1 era siempre cero)
- MAP6 se llama 'MAPE' en el savegame (no 'MAP6')

Fixes en Rust:
- Tile struct: agrega m6: u8 y m8: u16; from_ottd_binary lee v3 opcionales
- industry_sprite_for_gfx: ahora recibe u16 para soportar gfx >= 256
- main.rs: gfx = m5 | ((m6 >> 2) & 1) << 8 (9 bits correctos)
- map.rs doc: documenta nombres reales de chunks en el savegame

Los TrackBits de rail ya usaban 6 bits (m5 & 0x3F) correctamente.

Made-with: Cursor

+++ crates/openttdrs-client/src/sprites.rs
@@ -291,7 +291,9 @@ pub const INDUSTRY_GFX_DATA: [IndustryGfxSprite; 120] = [
 
 /// Devuelve los metadatos del sprite de industria para el gfx dado (byte m5).
 /// Retorna `None` si el gfx no tiene overlay de edificio (solo suelo) o está fuera del rango.
-pub fn industry_sprite_for_gfx(gfx: u8) -> Option<&'static IndustryGfxSprite> {
+/// Devuelve el sprite de industria para el índice gfx de 9 bits:
+/// `gfx = m5 | ((m6 >> 2) & 1) << 8`
+pub fn industry_sprite_for_gfx(gfx: u16) -> Option<&'static IndustryGfxSprite> {
     let entry = INDUSTRY_GFX_DATA.get(usize::from(gfx))?;
     if entry.sprite_id != 0 {
         Some(entry)
